further refactoring and cleaning up (a lot of changes, sorry -\_(oo)_/-)
