+++ dim/bigquery_client.py
@@ -3,8 +3,21 @@ import logging
 from typing import Any, Dict, Optional
 
 import attr
+from textwrap import dedent
 from google.cloud import bigquery
 from google.cloud.exceptions import NotFound
+from google.cloud.bigquery.job import WriteDisposition
+from google.cloud.bigquery.schema import SchemaField
+from google.cloud.bigquery.table import Table
+
+from dim.const import (
+    DESTINATION_DATASET,
+    DESTINATION_PROJECT,
+    PROCESSING_INFO_TABLE,
+    RUN_HISTORY_TABLE,
+    MUTED_ALERTS_TABLE,
+    MUTED_ALERTS_TABLE_NAME,
+)
 
 
 @attr.s(auto_attribs=True, slots=True)
@@ -74,3 +87,243 @@ class BigQueryClient:
             "total_bytes_billed": bq_query_job.total_bytes_billed,
             "total_bytes_processed": bq_query_job.total_bytes_processed,
         }
+
+
+def retrieve_dim_checks(project_id, dataset, table, run_uuid, failed_only=True):
+    sql = dedent(
+        f"""
+        SELECT
+            CONCAT(project_id, '.', dataset, '.', table) AS dataset,
+            dim_check_type,
+            dim_check_title,
+            dim_check_description,
+            FORMAT_DATE("%Y-%m-%d %H:%M:%S", actual_run_date) AS actual_run_date,  # noqa: E501
+            date_partition,
+            owner,
+            query_results,
+            dim_check_context,
+            run_id,
+            passed,
+        FROM `{RUN_HISTORY_TABLE}`
+        WHERE
+            project_id = '{project_id}'
+            AND dataset = '{dataset}'
+            AND table = '{table}'
+            AND run_id = '{run_uuid}'
+            {"AND NOT passed" if failed_only else ""}
+        """.replace(
+            "  # noqa: E501", ""
+        )
+    )
+
+    bigquery = BigQueryClient(project_id=DESTINATION_PROJECT, dataset=DESTINATION_DATASET)
+
+    job = bigquery.fetch_results(sql)
+
+    return job.result().to_dataframe()
+
+
+def insert_dim_processing_info(insert_data):
+    bigquery = BigQueryClient(project_id=DESTINATION_PROJECT, dataset=DESTINATION_DATASET)
+
+    processing_table = get_dim_processing_info_table()
+
+    result = bigquery.client.insert_rows(processing_table, insert_data)
+    run_id = insert_data[0]["run_id"]
+
+    if result:
+        logging.error(
+            "There has been a problem inserting processing \
+                info into: %s for run_id: %s"
+            % (str(processing_table), run_id)
+        )
+        logging.error(str(result))
+
+    logging.info(
+        "Processing info inserted into: %s for run_id: %s." % (PROCESSING_INFO_TABLE, run_id)
+    )
+
+
+def create_bq_table_if_not_exist(bq_client, table):
+    if not bq_client.bq_table_exists(table):
+        bq_client.create_table(table)
+
+
+def get_dim_processing_info_table():
+    schema = [
+        SchemaField(name="project_id", field_type="STRING"),
+        SchemaField(name="dataset", field_type="STRING"),
+        SchemaField(name="table", field_type="STRING"),
+        SchemaField(name="dim_check_type", field_type="STRING"),
+        SchemaField(name="dim_check_title", field_type="STRING"),
+        SchemaField(name="dim_check_description", field_type="STRING"),
+        SchemaField(name="date_partition", field_type="DATE"),
+        SchemaField(name="run_id", field_type="STRING"),
+        SchemaField(name="total_bytes_billed", field_type="INTEGER"),
+        SchemaField(name="total_bytes_processed", field_type="INTEGER"),
+    ]
+
+    table = Table(
+        PROCESSING_INFO_TABLE,
+        schema=schema,
+    )
+
+    bigquery = BigQueryClient(
+        project_id=DESTINATION_PROJECT,
+        dataset=DESTINATION_DATASET,
+    )
+
+    create_bq_table_if_not_exist(bigquery, table)
+
+    return table
+
+
+def get_muted_alerts_table():
+    schema = [
+        SchemaField(name="project_id", field_type="STRING"),
+        SchemaField(name="dataset", field_type="STRING"),
+        SchemaField(name="table", field_type="STRING"),
+        SchemaField(name="date_partition", field_type="DATE"),
+    ]
+
+    return Table(
+        MUTED_ALERTS_TABLE,
+        schema=schema,
+    )
+
+
+def is_alert_muted(
+    project_id: str,
+    dataset: str,
+    table: str,
+    date: str,
+) -> bool:
+    sql = dedent(
+        f"""
+        SELECT
+            COUNT(*) AS count,
+        FROM `{MUTED_ALERTS_TABLE}`
+        WHERE
+            project_id = '{project_id}'
+            AND dataset = '{dataset}'
+            AND table = '{table}'
+            AND DATE(date_partition) = DATE('{date}')
+        """
+    )
+
+    bigquery = BigQueryClient(
+        project_id=DESTINATION_PROJECT,
+        dataset=DESTINATION_DATASET,
+    )
+
+    bq_table = get_muted_alerts_table()
+
+    create_bq_table_if_not_exist(bigquery, bq_table)
+
+    job = bigquery.fetch_results(sql)
+    result_df = job.result().to_dataframe()
+
+    return bool(result_df["count"].iloc[0])
+
+
+def mute_alerts_for_date(
+    project_id: str,
+    dataset: str,
+    table: str,
+    date: Any,
+):
+    logging.info("Muting alerts for %s:%s.%s for date: %s" % (project_id, dataset, table, date))
+
+    bigquery = BigQueryClient(
+        project_id=DESTINATION_PROJECT,
+        dataset=DESTINATION_DATASET,
+    )
+
+    if is_alert_muted(project_id, dataset, table, date):
+        logging.info(
+            "Alerts already muted for %s:%s.%s for date: %s" % (project_id, dataset, table, date)
+        )
+        return
+
+    bq_table = get_muted_alerts_table()
+    create_bq_table_if_not_exist(bigquery, str(bq_table))
+
+    date_partition = date.date()
+
+    insert_data = [
+        {
+            "project_id": project_id,
+            "dataset": dataset,
+            "table": table,
+            "date_partition": date_partition,
+        }
+    ]
+
+    result = bigquery.client.insert_rows(bq_table, insert_data)
+
+    if not result:
+        logging.info(
+            "Alerts for %s:%s.%s for partition: %s have been muted."
+            % (project_id, dataset, table, date_partition)
+        )
+
+    # TODO: else: something went wrong, log error message
+
+
+def unmute_alerts_for_date(
+    project_id: str,
+    dataset: str,
+    table: str,
+    date: Any,
+):
+    logging.info("Unmuting alerts for %s:%s.%s for date: %s" % (project_id, dataset, table, date))
+
+    bigquery = BigQueryClient(
+        project_id=DESTINATION_PROJECT,
+        dataset=DESTINATION_DATASET,
+    )
+
+    bq_table = get_muted_alerts_table()
+    create_bq_table_if_not_exist(bigquery, bq_table)
+
+    if not is_alert_muted(project_id, dataset, table, date):
+        logging.info(
+            "Alerts not muted for %s:%s.%s for date: %s. \
+                Nothing needs to be done."
+            % (project_id, dataset, table, date)
+        )
+        return
+
+    date_partition = date.date()
+
+    delete_statement = dedent(
+        f"""
+        SELECT
+            *
+        FROM `{str(bq_table)}`
+        WHERE
+            NOT (
+                project_id = '{project_id}'
+                AND dataset = '{dataset}'
+                AND table = '{table}'
+                AND date_partition = '{date_partition}'
+            )
+        """
+    )
+
+    result, _ = bigquery.execute(
+        delete_statement,
+        dataset=DESTINATION_DATASET,
+        destination_table=MUTED_ALERTS_TABLE_NAME,
+        write_disposition=WriteDisposition.WRITE_TRUNCATE,
+        schema_update_options=None,
+    )
+
+    if not result:
+        # TODO: else: something went wrong, log error message
+        pass
+
+    logging.info(
+        "Alerts for %s:%s.%s for partition: %s has been unmuted."
+        % (project_id, dataset, table, date_partition)
+    )
