+++ dim/cli.py
@@ -7,7 +7,8 @@ from dim.app import run_check
 from dim.const import INPUT_DATE_FORMAT, LOGGING_LEVEL, SOURCE_PROJECT
 from dim.error import CmdDateInfoNotProvidedException, DateRangeException, DimConfigError
 from dim.models.dim_config import DimConfig
-from dim.utils import mute_alerts_for_date, read_config, unmute_alerts_for_date
+from dim.utils import read_config
+from dim.bigquery_client import mute_alerts_for_date, unmute_alerts_for_date
 
 logging.basicConfig(level=LOGGING_LEVEL)
