+++ dim/models/dim_check_type/column_sum_not_zero.py
@@ -9,12 +9,7 @@ class ColumnSumNotZero(Base):
         project_id,
         dataset,
         table,
+        dim_check_title="",
         dim_check_description="",
     ):
-        super().__init__(project_id, dataset, table, dim_check_type="column_sum_not_zero", dim_check_description=dim_check_description)
-
-    def generate_test_sql(self, *, params: Dict[Any, Any]) -> str:
-        return super().generate_test_sql(params=params)
-
-    def execute_test_sql(self, sql: str) -> Any:
-        return super().execute_test_sql(sql=sql)
+        super().__init__(project_id, dataset, table, dim_check_type="column_sum_not_zero", dim_check_title=dim_check_title, dim_check_description=dim_check_description)
