+++ dim/models/dim_check_type/templates/compare_row_count_to_table.sql.jinja
@@ -20,7 +20,8 @@ SELECT
     '{{ params.tier }}' AS tier,
     DATE('{{ params.partition }}') AS date_partition,
     '{{ dim_check_type }}' AS dim_check_type,
-  '{{ dim_check_description }}' AS dim_check_description,
+    '{{ dim_check_title }}' AS dim_check_title,
+    '{{ dim_check_description }}' AS dim_check_description,
     IF({% if params.acceptable_delta %}ABS(row_count - {{ other_table[2] }}_row_count) <= {{ params.acceptable_delta }}{% else%}row_count = {{ other_table[2] }}_row_count{% endif %}, True, False) AS passed,
     '{{ params.owner }}' AS owner,
     TO_JSON_STRING(CTE) AS query_results,
