+++ dim/models/dim_check_type/templates/custom_sql_metric.sql.jinja
@@ -8,7 +8,8 @@ SELECT
     '{{ table }}' AS `table`,
     '{{ params.tier }}' AS tier,
     '{{ dim_check_type }}' AS dim_check_type,
-  '{{ dim_check_description }}' AS dim_check_description,
+    '{{ dim_check_title }}' AS dim_check_title,
+    '{{ dim_check_description }}' AS dim_check_description,
     DATE('{{ params.partition }}') AS date_partition,
     IF({{ params.condition }}, True, False) AS passed,
     '{{ params.owner }}' AS owner,
