+++ dim/models/dim_check_type/templates/table_row_count.sql.jinja
@@ -12,7 +12,8 @@ SELECT
     '{{ params.tier }}' AS tier,
     DATE('{{ params.partition }}') AS date_partition,
     '{{ dim_check_type }}' AS dim_check_type,
-  '{{ dim_check_description }}' AS dim_check_description,
+    '{{ dim_check_title }}' AS dim_check_title,
+    '{{ dim_check_description }}' AS dim_check_description,
     IF(row_count {{ params.condition }}, True, False) AS passed,
     '{{ params.owner }}' AS owner,
     TO_JSON_STRING(CTE) AS query_results,
