+++ dim/models/dim_config.py
@@ -69,10 +69,6 @@ class DimCheckParams:
     acceptable_delta: Optional[int] = attr.ib(None)
     metric_1: Optional[Metric] = attr.ib(None)
     metric_2: Optional[Metric] = attr.ib(None)
-    # metric_name_1: Optional[str] = attr.ib(None)
-    # metric_query_1: Optional[str] = attr.ib(None)
-    # metric_name_2: Optional[str] = attr.ib(None)
-    # metric_query_2: Optional[str] = attr.ib(None)
 
 
 @attr.s(auto_attribs=True)
@@ -81,6 +77,7 @@ class DimCheck:
 
     type: str = attr.ib()
     params: DimCheckParams
+    title: Optional[str] = attr.ib(None)
     description: Optional[str] = attr.ib(None)
 
     @type.validator
