+++ dim_checks/mozdata/firefox_desktop/active_users_aggregates/dim_checks.yaml
@@ -21,6 +21,7 @@ dim_config:
           - country
         condition: "= 2"
     - type: custom_compare_metrics
+      title: DAU validation
       description: Validating that DAU sum in Active Users Aggregates is the same as unique client_id count in the live table.
       params:
         condition: dau_sum = distinct_client_count
@@ -55,6 +56,7 @@ dim_config:
             FULL OUTER JOIN overactive USING(client_id)
             WHERE overactive.client_id IS NULL
     - type: custom_compare_metrics
+      name: QDAU validation
       description: Validating that QDAU sum in Active Users Aggregates is as expected based on the live table.
       params:
         condition: qdau_sum = live_table_qdau_count
